+++ src/pull.ts
@@ -560,6 +560,8 @@ async function pullFilesFromWithProgress(progress: deploy_helpers.ProgressContex
                 succeeded: [],
             };
 
+            const OVERALL_WATCH = deploy_helpers.startWatch();
+
             // "finished button"
             await ME.invokeForFinishedButton(
                 deploy_contracts.DeployOperation.Pull,
@@ -593,15 +595,11 @@ async function pullFilesFromWithProgress(progress: deploy_helpers.ProgressContex
                         btn.color = new vscode.ThemeColor(color);
                         btn.text = `${icon} ` +
                                    `[${NOW.format( ME.t('time.timeWithSeconds') )}] ` + 
-                                   ME.t('pull.finishedButton.text');
+                                   ME.t('pull.finishedButton.text',
+                                        OVERALL_WATCH.stop());
                         btn.tooltip = ME.t('pull.finishedButton.tooltip');
 
-                        btn.show();
-
-                        ME.setTimeoutForFinishedButton(
-                            deploy_contracts.DeployOperation.Pull,
-                            (b) => b.hide()
-                        );
+                        ME.showFinishedButton(deploy_contracts.DeployOperation.Pull);
                     }
                 );
             };